Before:
    
    [2025-09-10 20:33:56] Starting kcidb validate+upgrade benchmark: 1000 files (5.95 MB)
    [2025-09-10 20:33:57] Progress: 200/1000 | 1.60s elapsed | 125.0 files/s | 0.43 MB (0.27 MB/s)
    [2025-09-10 20:33:59] Progress: 400/1000 | 3.21s elapsed | 124.5 files/s | 1.84 MB (0.57 MB/s)
    [2025-09-10 20:34:00] Progress: 600/1000 | 4.80s elapsed | 125.0 files/s | 2.27 MB (0.47 MB/s)
    [2025-09-10 20:34:02] Progress: 800/1000 | 6.43s elapsed | 124.4 files/s | 5.70 MB (0.89 MB/s)
    [2025-09-10 20:34:04] Progress: 1000/1000 | 8.02s elapsed | 124.6 files/s | 5.95 MB (0.74 MB/s)
    [2025-09-10 20:34:04] Benchmark complete
    Processed: 1000 files, 5.95 MB in 8.02s | 124.6 files/s | 0.74 MB/s
    Avg per file (s): read=0.0000 parse=0.0000 validate=0.0079 upgrade=0.0000
    
    Let's make kcidb-io better:
      - Added per-class cached validator in kcidb_io/schema/abstract.py.
      - validate_exactly() now:
          - Lazily caches a draft7 format_checker.
          - Lazily compiles and caches a jsonschema.Draft7Validator for the class’ schema.
          - Calls cls._validator.validate(data) instead of re-creating a validator each call.
      - Ensured the cache is per-class (not inherited) by checking cls.__dict__.
    
    After:
    
    [2025-09-10 20:38:57] Starting kcidb validate+upgrade benchmark: 1000 files (5.95 MB)
    [2025-09-10 20:38:57] Progress: 200/1000 | 0.07s elapsed | 2700.9 files/s | 0.43 MB (5.77 MB/s)
    [2025-09-10 20:38:57] Progress: 400/1000 | 0.14s elapsed | 2789.8 files/s | 1.84 MB (12.86 MB/s)
    [2025-09-10 20:38:57] Progress: 600/1000 | 0.20s elapsed | 3026.8 files/s | 2.27 MB (11.46 MB/s)
    [2025-09-10 20:38:57] Progress: 800/1000 | 0.31s elapsed | 2594.0 files/s | 5.70 MB (18.49 MB/s)
    [2025-09-10 20:38:57] Progress: 1000/1000 | 0.37s elapsed | 2732.3 files/s | 5.95 MB (16.27 MB/s)
    [2025-09-10 20:38:57] Benchmark complete
    Processed: 1000 files, 5.95 MB in 0.37s | 2732.2 files/s | 16.27 MB/s
    Avg per file (s): read=0.0000 parse=0.0000 validate=0.0003 upgrade=0.0000